+++ cdisc_rules_engine/sql_operations/date_operation.py
@@ -24,12 +24,12 @@ class SqlDateOperation(SqlBaseOperation):
                             CASE
                                 WHEN {self.function}(CASE
                                     WHEN {column_id} IS NULL OR {column_id} = '' THEN NULL
-                                    ELSE {column_id}::date
+                                    ELSE {column_id}::timestamp
                                 END) IS NULL THEN ''
                                 ELSE TO_CHAR({self.function}(CASE
                                     WHEN {column_id} IS NULL OR {column_id} = '' THEN NULL
-                                    ELSE {column_id}::date
-                                END), 'YYYY-MM-DD')
+                                    ELSE {column_id}::timestamp
+                                END), 'YYYY-MM-DD"T"HH24:MI:SS')
                             END, ''
                         ) AS value
                         FROM {dataset_id}"""
@@ -61,12 +61,12 @@ class SqlDateOperation(SqlBaseOperation):
                             CASE
                                 WHEN {self.function}(CASE
                                     WHEN {column_id} IS NULL OR {column_id} = '' THEN NULL
-                                    ELSE {column_id}::date
+                                    ELSE {column_id}::timestamp
                                 END) IS NULL THEN ''
                                 ELSE TO_CHAR({self.function}(CASE
                                     WHEN {column_id} IS NULL OR {column_id} = '' THEN NULL
-                                    ELSE {column_id}::date
-                                END), 'YYYY-MM-DD')
+                                    ELSE {column_id}::timestamp
+                                END), 'YYYY-MM-DD"T"HH24:MI:SS')
                             END, ''
                         ) AS value
                         FROM {dataset_id}
